VC warning fixes
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@4718 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/stream.cpp
@@ -393,8 +393,8 @@ struct filestream : stream
     bool end() { return feof(file)!=0; }
     long tell() { return ftell(file); }
     bool seek(long offset, int whence) { return fseek(file, offset, whence) >= 0; }
-    int read(void *buf, int len) { return fread(buf, 1, len, file); }
-    int write(const void *buf, int len) { return fwrite(buf, 1, len, file); }
+    int read(void *buf, int len) { return (int)fread(buf, 1, len, file); }
+    int write(const void *buf, int len) { return (int)fwrite(buf, 1, len, file); }
     int getchar() { return fgetc(file); }
     bool putchar(int c) { return fputc(c, file)!=EOF; }
     bool getline(char *str, int len) { return fgets(str, len, file)!=NULL; }
